configured project for deployment to test network
