+++ hardhat.config.js
@@ -1,0 +1,9 @@
+require("@nomiclabs/hardhat-waffle");
+/**
+ * @type import('hardhat/config').HardhatUserConfig
+ */
+module.exports = {
+  solidity: {
+    version: "0.8.0"
+  }
+};
