+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "FrenlyDogs",
+  "version": "1.0.0",
+  "description": "WIP NFT contract",
+  "main": "FND.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mkUltraVictim/FrenlyDogs.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/mkUltraVictim/FrenlyDogs/issues"
+  },
+  "homepage": "https://github.com/mkUltraVictim/FrenlyDogs#readme",
+  "devDependencies": {
+    "@nomiclabs/hardhat-ethers": "^2.0.3",
+    "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "chai": "^4.3.4",
+    "ethereum-waffle": "^3.4.0",
+    "ethers": "^5.5.2",
+    "hardhat": "^2.7.1"
+  },
+  "dependencies": {
+    "@openzeppelin/contracts": "^4.4.0"
+  }
+}
